requirements.txt
phidata
streamlit
openai
yfinance
duckduckgo-search